tests/requirements-test.txt
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.1
pytest-asyncio>=0.21.1
pytest-mock>=3.11.1
coverage>=7.3.0
requests-mock>=1.11.0
responses>=0.23.8
freezegun>=1.2.2
pyyaml>=6.0.1 